app/requirements.txt
eventlet==0.40.3
Flask==3.1.2
Flask_SocketIO==5.5.1
kubernetes==34.1.0
python-dotenv==1.2.1
redis==7.0.1
